Add files via upload
We presents CaroliO, an ontology for Caroli's disease and syndrome, aimed at enhancing the diagnosis and treatment of these rare liver conditions. Our work addresses the pressing need for structured knowledge resources in the field of rare diseases, which often present significant diagnostic and treatment challenges due to limited information.